Builder character implemented
Quite a lot of changes.
Added a "blueprint" system where the decision to build is started and
the building is placed down. Towers do not become active until they have
reached maximum health.
I had a lot of trouble getting the collision to work with this, along
with the green/blue visualisation.